+++ src/main/java/me/pajic/affogato_core/CompatFlags.java
@@ -3,6 +3,7 @@ package me.pajic.affogato_core;
 import net.fabricmc.loader.api.FabricLoader;
 
 public class CompatFlags {
+    public static final boolean IRIS_LOADED = FabricLoader.getInstance().isModLoaded("iris");
     public static final boolean FD_LOADED = FabricLoader.getInstance().isModLoaded("farmersdelight");
     public static final boolean ITEMSWAPPER_LOADED = FabricLoader.getInstance().isModLoaded("itemswapper");
     public static final boolean SERENE_WILD_LOADED = FabricLoader.getInstance().isModLoaded("serenewild");